+++ src/main/java/com/supercoding/shoppingmallbackend/dto/response/profile/RechargeResponse.java
@@ -1,0 +1,11 @@
+package com.supercoding.shoppingmallbackend.dto.response.profile;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@AllArgsConstructor
+public class RechargeResponse {
+    private Long rechargeMoney;
+}
